refactor: extract useDebouncedValue to @pops/ui, improve debounce test, fix condition docs
- Add useDebouncedValue to packages/ui/src/lib/useDebounce.ts and
  export from @pops/ui to eliminate 4 local copies of the same hook
- Remove inline useDebouncedValue from ItemsPage, SearchPage,
  LibraryPage, ComparisonHistoryPage; import from @pops/ui instead
- Improve debounce test: assert search term absent immediately after
  change, still absent at 299ms, present at 300ms exactly
- Align us-03 condition options doc with actual app vocabulary:
  Excellent/Good/Fair/Poor (not new/good/fair/poor/broken)
