+++ packages/app-media/src/pages/SearchPage.tsx
@@ -17,7 +17,7 @@ import { toast } from 'sonner';
  * "Watched + Library" buttons that add the item then trigger the secondary
  * action in a single click.
  */
-import { Button, Skeleton, Tabs, TabsList, TabsTrigger, TextInput } from '@pops/ui';
+import { Button, Skeleton, Tabs, TabsList, TabsTrigger, TextInput, useDebouncedValue } from '@pops/ui';
 
 import {
   buildPosterUrl,
@@ -50,19 +50,7 @@ interface TvSearchResult {
   year: string | null;
 }
 
-/** Hook: debounce a string value by `delay` ms. */
-function useDebouncedValue(value: string, delay: number): string {
-  const [debounced, setDebounced] = useState(value);
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setDebounced(value);
-    }, delay);
-    return () => {
-      clearTimeout(timer);
-    };
-  }, [value, delay]);
-  return debounced;
-}
+
 
 export function SearchPage() {
   const [searchParams, setSearchParams] = useSearchParams();
