+++ packages/ui/src/lib/useDebounce.ts
@@ -1,0 +1,11 @@
+import { useEffect, useState } from 'react';
+
+/** Debounce a string value by `delay` ms. Returns the debounced copy. */
+export function useDebouncedValue(value: string, delay: number): string {
+  const [debounced, setDebounced] = useState(value);
+  useEffect(() => {
+    const timer = setTimeout(() => setDebounced(value), delay);
+    return () => clearTimeout(timer);
+  }, [value, delay]);
+  return debounced;
+}
